Split up array manipulations to separate file
