+++ rustynum-rs/src/num_array/mod.rs
@@ -2,6 +2,7 @@ mod array_struct;
 mod constructors;
 mod impl_clone_from;
 pub mod linalg;
+mod manipulation;
 pub mod operations;
 
 pub use array_struct::{NumArray, NumArrayF32, NumArrayF64, NumArrayI32, NumArrayI64, NumArrayU8};
